feat(azure): support azure blob container storage
 - add azure implementation, and register with scheme "azure"
 - update test suite to use azure client when "azure" adapter is requested
 - add `pathy[azure]` extras to setup.py
 - remove "list_buckets" from BucketClient classes

BREAKING CHANGE: This removes an internal bit of code that allowed for enumerating buckets in certain situations. The API was impossible to reach without going indirectly through the glob functionality, and it's not clear that the codepaths were ever reached outside of specific unit testing situations. If there's an explicit need for listing buckets, we can add a top-level API for it.