refactor(Pathy): integrate some of pathlib_abc
 - dirty first pass, validating against the FS adapter test suite.
 - 38 fail / 92 passed so far
 - test_pathy, test_pure_pathy, test_cli left for FS adapter

BREAKING CHANGE: Pathy.key returns a str rather than a Pathy instance

There was nowhere in the code where I wasn't casting the .key access to string. This removes all those casts and saves a little effort. It's possible that some user expects .key to be a Path still, so add a breaking change.